requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
google-generativeai==0.8.5
jinja2==3.1.2
